refactor: make deploy one operation that forwards its phases' events
`commands/deploy.rs` was 973 lines of orchestration, opening seven separate
renderers — build, create, env vars, settings, Candid, install, sync — and
awaiting each before starting the next. That sequencing was not a style
choice: it was the mechanism keeping the `info!("Installing canisters:")`
headings from tearing through a live progress view. It did not fully hold,
either — the `println!("Created canister … with ID …")` inside the create
scope raced the MultiProgress it was printing over.

Deploy is now an operation. It takes one `&Reporter` and runs the whole thing
as a single task tree: each phase is a `Shape::Group` task, and the reporter
handed to `build_many`, `install_many`, `sync_many` and the rest is scoped to
that phase, so their per-canister tasks nest under it. None of those
operations changed — they still take a `&Reporter` and still start what they
think are top-level tasks. That is the property that makes this worth doing:
composing an operation costs nothing at the operations it composes.

Consequences:

  - The phase headings are events now, ordered on the same stream as the bars
    they title, so nothing has to be awaited to keep them in place. The
    renderer closes the live view before a top-level heading, which reproduces
    what tearing down a per-phase renderer used to do.
  - Nothing in the operation writes to the terminal. `DeployReport` carries the
    canisters it created and the command prints them, so stdout stays in the
    command layer next to `--json`. The println/MultiProgress race is gone
    structurally rather than by careful placement.
  - `commands/deploy.rs` drops to 476 lines: argument handling, the
    `http_request` probe, URL printing, and the JSON shape. The readiness
    probe (`wait_until_serving_queries`) moves down with the sync phase it
    guards.
  - The settings phase gained the heading it never had; its bars used to
    appear under `Setting environment variables:`.

`resolve_targets` is split out so the command can decide there is nothing to
do, and reject `--args` with more than one canister, before starting a run.

All 24 deploy tests pass unmodified. Next step, and the reason this is shaped
this way: PR #713's deploy library needs exactly this seam — it currently
threads `Option<&dyn StepProgress>` through `PluginExecutor`, `ScriptRunner`
and the fetch path to keep terminal output out of a library that has to run
inside a canister. A composite operation reporting generic events is that
seam, and it already works for a host with no terminal at all.
